fix timestamp when time travel tests
