+++ contracts/HarmonyBridge.sol
@@ -181,7 +181,7 @@ contract Bridge is Ownable {
 
     function removeValidator(address removedValidator) public onlyOwner() {
         require(
-            currentValidatorsCount -1 >= signatureThreshold,
+            currentValidatorsCount - 1 >= signatureThreshold,
             "There are no validators now!"
         );
         validators[removedValidator] = false;
@@ -214,7 +214,7 @@ contract Bridge is Ownable {
         public
         onlyOwner()
     {
-        require(tokenDailyLimit !=0, "Invalid value tokenDailyLimit");
+        require(tokenDailyLimit != 0, "Invalid value tokenDailyLimit");
         tokens[newToken] = true;
         dailyLimit[newToken] = tokenDailyLimit;
         dailyLimitSetTime[newToken] = block.timestamp;
@@ -267,9 +267,9 @@ contract Bridge is Ownable {
 
     function isTimeNotExpired(uint256 txTime) private view returns (bool) {
         if (block.timestamp.sub(txTime) < durationBeforeExpirationTime) {
-            return false;
-        } else {
             return true;
+        } else {
+            return false;
         }
     }
 
@@ -319,9 +319,7 @@ contract Bridge is Ownable {
             assetDailyLimit > 0,
             "Can't transfer asset without daily limit"
         );
-
         updateDailyLimit(transferInfo.asset);
-
         require(
             transferInfo.amount.add(dailySpend[transferInfo.asset]) <=
                 assetDailyLimit,
@@ -331,7 +329,6 @@ contract Bridge is Ownable {
         dailySpend[transferInfo.asset] = dailySpend[transferInfo.asset].add(
             transferInfo.amount
         );
-
         if (transferInfo.asset == address(0)) {
             uint256 amountToSend = transferInfo.amount.sub(
                 transferInfo.amount.mul(fee).div(100)
