+++ test/bridge.js
@@ -1,4 +1,3 @@
-// const { count, assert } = require('console');
 const keccak256 = require("keccak256");
 const ethSig = require("nano-ethereum-signer");
 const assert = require("assert");
@@ -6,9 +5,9 @@ const assert = require("assert");
 const HarmonyBridge = artifacts.require("Bridge");
 const EdgewareToken = artifacts.require("EdgewareToken");
 
-const currentTime = (Date.now() / 1000).toFixed();
+const currentTime = (Date.now()/1000).toFixed();
 
-contract("Bridge", async (accounts) => {
+contract("HarmonyBridge", async (accounts) => {
   contract("Change Bridge parameters", async (accounts) => {
     let bridgeContract, edgewareToken;
 
@@ -187,226 +186,7 @@ contract("Bridge", async (accounts) => {
     });
   });
 
-  contract("Submit signatures", async (accounts) => {
-    let bridgeContract;
-    let edgewareToken;
-    before(async () => {
-      bridgeContract = await HarmonyBridge.deployed();
-      edgewareToken = await EdgewareToken.deployed();
 
-      let keys = Object.keys(validators);
-      for (let i = 0; i < keys.length; i++) {
-        await bridgeContract.addValidator(keys[i], { from: accounts[0] });
-      }
-      await bridgeContract.addWorker(accounts[1], { from: accounts[0] });
-      await bridgeContract.addToken(
-        edgewareToken.address,
-        10,
-        { from: accounts[0] }
-      );
-    });
-
-    it("Send swap request", async () => {
-      const startBalance = 5;
-      let message = getSwapMessage(accounts[3]);
-      await bridgeContract.send(startBalance * 1 ** 18, { from: accounts[2] });
-
-      let balanceBefore = await web3.eth.getBalance(bridgeContract.address);
-      assert.strictEqual(parseInt(balanceBefore), startBalance);
-      let signatures = signSwapMessage(hashMessage(message), 4);
-
-      await bridgeContract.setFee(1, { from: accounts[0] });
-      await bridgeContract.requestSwap(message, signatures, {
-        from: accounts[1],
-      });
-      const expectedBalance = startBalance - message.amount;
-      let balanceAfterTest = await web3.eth.getBalance(bridgeContract.address);
-      assert.strictEqual(parseInt(balanceAfterTest), expectedBalance);
-    });
-
-    it("Send token swap", async () => {
-      const startBalance = 5;
-      let message = getSwapMessage(accounts[2]);
-      message.asset = edgewareToken.address;
-      await bridgeContract.send(startBalance * 1 ** 18, { from: accounts[2] });
-      let signatures = signSwapMessage(hashMessage(message), 4);
-      await edgewareToken.setBridgeAddress(bridgeContract.address, {from: accounts[0]});
-      const {receipt} = await bridgeContract.requestSwap(message, signatures, {
-        from: accounts[1],
-        gasValue: 200000,
-      });
-      assert.strictEqual(Number(receipt.rawLogs[0].data) , message.amount);
-    });
-
-    it("Send swap request with wrong count of signatures", async () => {
-      const startBalance = 5;
-      let message = getSwapMessage(accounts[2]);
-      let signatures = signSwapMessage(hashMessage(message), 3);
-      await bridgeContract.send(startBalance * 1 ** 18, { from: accounts[2] });
-      try {
-        await bridgeContract.requestSwap(message, signatures, {
-          from: accounts[1],
-        });
-        assert.fail();
-      } catch (error) {
-        assert(
-          error
-            .toString()
-            .includes("Wrong count of signatures to make transfer"),
-          error.toString()
-        );
-      }
-    });
-
-    it("Send swap request with one signature not from Validator", async () => {
-      const startBalance = 5;
-      let message = getSwapMessage(accounts[2]);
-      await bridgeContract.send(startBalance * 1 ** 18, { from: accounts[2] });
-      let signatures = signSwapMessage(hashMessage(message), 5);
-      signatures.splice(0, 1);
-      try {
-        await bridgeContract.requestSwap(message, signatures, {
-          from: accounts[1],
-        });
-      } catch (error) {
-        assert(
-          error.toString().includes("Signatures verification is failed"),
-          error.toString()
-        );
-      }
-    });
-
-    it("Send swap request with one signature of wrong data", async () => {
-      const startBalance = 5;
-      await bridgeContract.send(startBalance * 1 ** 18, { from: accounts[2] });
-      let message = getSwapMessage(accounts[2]);
-      let signatures = signSwapMessage(hashMessage(message), 4);
-      message.amount = 1;
-      let wrong_signature = signSwapMessage(hashMessage(message), 1);
-      signatures.splice(0, 1);
-      signatures.push(wrong_signature[0]);
-      try {
-        await bridgeContract.requestSwap(message, signatures, {
-          from: accounts[1],
-        });
-        assert.fail();
-      } catch (error) {
-        assert(
-          error.toString().includes("Signatures verification is failed"),
-          error.toString()
-        );
-      }
-    });
-
-    it("Send swap request signed only one Validator", async () => {
-      const startBalance = 5;
-      await bridgeContract.send(startBalance * 1 ** 18, { from: accounts[2] });
-      let message = getSwapMessage(accounts[2]);
-      let signatures = [];
-      for (let i = 0; i < 4; i++) {
-        let s = signSwapMessage(hashMessage(message), 1);
-        signatures.push(s[0]);
-      }
-      try {
-        await bridgeContract.requestSwap(message, signatures, {
-          from: accounts[1],
-        });
-        assert.fail();
-      } catch (error) {
-        assert(
-          error.toString().includes("Signatures verification is failed"),
-          error.toString()
-        );
-      }
-    });
-
-    it("Send token swap and reached daily limit", async () => {
-      const startBalance = 20;
-      await bridgeContract.send(startBalance * 1 ** 18, { from: accounts[2] });
-      let message = getSwapMessage(accounts[2]);
-
-      message.asset = edgewareToken.address;
-      message.amount = 8;
-      await edgewareToken.setBridgeAddress(bridgeContract.address, {from: accounts[0]});
-      let signatures = signSwapMessage(hashMessage(message), 4);
-      try {
-      await bridgeContract.requestSwap(message, signatures, {
-        from: accounts[1],
-      });
-      assert.fail();
-    } catch (error) {
-      assert(
-        error.toString().includes("Daily limit has already reached for this asset"),
-        error.toString()
-      );
-    }
-  });
-});
-
-  contract("Submit signatures with time travel", async (accounts) => {
-    it("Send token swap and check updating daily limit", async () => {
-      let ctrObj = await HarmonyBridge.deployed();
-      let edgewareToken = await EdgewareToken.deployed();
-      const startBalance = 20;
-      await ctrObj.send(startBalance * 1 ** 18, { from: accounts[2] });
-      let keys = Object.keys(validators);
-      for (let i = 0; i < keys.length - 1; i++) {
-        await ctrObj.addValidator(keys[i]);
-      }
-      await ctrObj.addWorker(accounts[1]);
-      await ctrObj.addToken(edgewareToken.address, 10, {from: accounts[0]});
-      await edgewareToken.setBridgeAddress(ctrObj.address, {from: accounts[0]});
-     //  await ctrObj.setNewDurationBeforeExpirationTime((Date.now() / 1000).toFixed());
-
-      let message = getSwapMessage(accounts[2]);
-      message.asset = edgewareToken.address;
-      message.amount = 1;
-      let signatures = signSwapMessage(hashMessage(message), 4);
-      const  {receipt} = await ctrObj.requestSwap(message, signatures, {
-        from: accounts[1],
-      });
-      assert.strictEqual(Number(receipt.rawLogs[0].data) , message.amount);
-      await advanceTime(86401);
-      message.amount = 1;
-      message.timestamp = (Date.now() / 1000).toFixed();
-      await ctrObj.addToken(edgewareToken.address, 10, {from: accounts[0]});
-      signatures = signSwapMessage(hashMessage(message), 4);
-      const res = await ctrObj.requestSwap(message, signatures, {
-        from: accounts[1],
-      });
-      let secondSpent = await ctrObj.dailySpend(
-        edgewareToken.address
-      );
-      assert.strictEqual(Number(res.receipt.rawLogs[0].data) , message.amount);
-      assert.strictEqual(secondSpent.toString(), "2");
-    });
-
-    it("Send token swap with expiration time", async () => {
-      let ctrObj = await HarmonyBridge.deployed();
-      let edgewareToken = await EdgewareToken.deployed();
-      let keys = Object.keys(validators);
-      for (let i = 0; i < keys.length - 1; i++) {
-        await ctrObj.addValidator(keys[i]);
-      }
-      await ctrObj.addWorker(accounts[1]);
-      await ctrObj.addToken(edgewareToken.address, 10);
-      await edgewareToken.setBridgeAddress(ctrObj.address, {from: accounts[0]});
-      let message = getSwapMessage(accounts[2]);
-      message.asset = edgewareToken.address;
-      let signatures = signSwapMessage(hashMessage(message), 4);
-      await advanceTime(86402);
-      try {
-        await ctrObj.requestSwap(message, signatures, {
-          from: accounts[1],
-        });
-      } catch (error) {
-        assert(
-          error.toString().includes("Transaction can't be sent because of expiration time"),
-          error.toString()
-        );
-      }
-    });
-  });
 });
 
 const validators = {
@@ -461,22 +241,3 @@ function signSwapMessage(message, countOfSignatures) {
   }
   return signatures;
 }
-
-advanceTime = (time) => {
-  return new Promise((resolve, reject) => {
-    web3.currentProvider.send(
-      {
-        jsonrpc: "2.0",
-        method: "evm_increaseTime",
-        params: [time],
-        id: new Date().getTime(),
-      },
-      (err, result) => {
-        if (err) {
-          return reject(err);
-        }
-        return resolve(result);
-      }
-    );
-  });
-};
